+++ htdocs/cabinetmed/admin/admin.php
@@ -122,7 +122,7 @@ $h++;
 print '<form name="cabinetmed" action="'.$_SERVER["PHP_SELF"].'" method="POST">';
 print '<input type="hidden" name="action" value="update">';
 
-dol_fiche_head($head, 'tabsetup', '');
+dol_fiche_head($head, 'tabsetup', '', ((float) DOL_VERSION < 7.0 ? 0 : -1));
 
 $var=true;